requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
python-multipart==0.0.20
h11==0.16.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
twilio==9.7.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
redis==5.2.1
asyncpg==0.30.0
pytz==2024.1

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.25.2

# Document processing for knowledge base
PyPDF2==3.0.1
python-docx==1.1.2
